requirements-dev.txt
# Development dependencies
# Install with: pip install -r requirements-dev.txt

# Include production requirements
-r requirements.txt

# Code formatting
black==24.1.1

# Type checking
mypy==1.8.0
django-stubs==4.2.7
types-requests==2.32.4.20250809

# Pre-commit hooks
pre-commit==3.6.0

# Testing
pytest==8.0.0
pytest-django==4.8.0
pytest-cov==4.1.0
coverage==7.4.1 